ccd-ai-agent/requirements.txt
google-adk
google-cloud-storage
deprecated